# minimal jujuc 1867021