--- a/acceptancetests/repository/charms/fill-logs/actions/fill-machine
+++ b/acceptancetests/repository/charms/fill-logs/actions/fill-machine
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+go run actions/actions.go fill-machine