--- a/acceptancetests/repository/charms/fill-logs/functions/fill-machine
+++ b/acceptancetests/repository/charms/fill-logs/functions/fill-machine
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-go run actions/actions.go fill-machine