--- a/cmd/jujuc/main.go
+++ b/cmd/jujuc/main.go
@@ -1,0 +1,153 @@
+// Copyright 2012-2014 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package main
+
+import (
+	"os"
+	"path/filepath"
+	"runtime"
+
+	"github.com/juju/cmd"
+	"github.com/juju/errors"
+	"github.com/juju/loggo"
+	"github.com/juju/utils/exec"
+
+	"github.com/juju/juju/juju/sockets"
+)
+
+var logger = loggo.GetLogger("juju.cmd.jujud")
+
+
+const (
+	// exit_err is the value that is returned when the user has run juju in an invalid way.
+	exit_err = 2
+	// exit_panic is the value that is returned when we exit due to an unhandled panic.
+	exit_panic = 3
+)
+
+
+// copied from github.com/juju/juju/worker/uniter/runner/jujuc/server.go
+
+type Request struct {
+    ContextId   string
+    Dir         string
+    CommandName string
+    Args        []string
+
+    // StdinSet indicates whether or not the client supplied stdin. This is
+    // necessary as Stdin will be nil if the client supplied stdin but it
+    // is empty.
+    StdinSet bool
+    Stdin    []byte
+
+    Token string
+}
+
+func getenv(name string) (string, error) {
+	value := os.Getenv(name)
+	if value == "" {
+		return "", errors.Errorf("%s not set", name)
+	}
+	return value, nil
+}
+
+func getwd() (string, error) {
+	dir, err := os.Getwd()
+	if err != nil {
+		return "", err
+	}
+	abs, err := filepath.Abs(dir)
+	if err != nil {
+		return "", err
+	}
+	return abs, nil
+}
+
+func getSocket() (sockets.Socket, error) {
+	var err error
+	socket := sockets.Socket{}
+	socket.Address, err = getenv("JUJU_AGENT_SOCKET_ADDRESS")
+	if err != nil {
+		return sockets.Socket{}, err
+	}
+	socket.Network, err = getenv("JUJU_AGENT_SOCKET_NETWORK")
+	if err != nil {
+		return sockets.Socket{}, err
+	}
+
+        return socket, nil
+}
+
+// hookToolMain uses JUJU_CONTEXT_ID and JUJU_AGENT_SOCKET_ADDRESS to ask a running unit agent
+// to execute a Command on our behalf. Individual commands should be exposed
+// by symlinking the command name to this executable.
+func hookToolMain(commandName string, ctx *cmd.Context, args []string) (code int, err error) {
+	code = 1
+	contextID, err := getenv("JUJU_CONTEXT_ID")
+	if err != nil {
+		return
+	}
+	dir, err := getwd()
+	if err != nil {
+		return
+	}
+	req := Request{
+		ContextId:   contextID,
+		Dir:         dir,
+		CommandName: commandName,
+		Args:        args[1:],
+		Token:       os.Getenv("JUJU_AGENT_TOKEN"),
+	}
+	socket, err := getSocket()
+	if err != nil {
+		return
+	}
+	client, err := sockets.Dial(socket)
+	if err != nil {
+		return code, err
+	}
+	defer client.Close()
+	var resp exec.ExecResponse
+	err = client.Call("Jujuc.Main", req, &resp)
+	if err != nil {
+		return
+	}
+	os.Stdout.Write(resp.Stdout)
+	os.Stderr.Write(resp.Stderr)
+	return resp.Code, nil
+}
+
+// MainWrapper exists to preserve test functionality.
+// On windows we need to catch the return code from main for
+// service functionality purposes, but on unix we can just os.Exit
+func MainWrapper(args []string) {
+	os.Exit(Main(args))
+}
+
+// Main is not redundant with main(), because it provides an entry point
+// for testing with arbitrary command line arguments.
+func Main(args []string) int {
+	defer func() {
+		if r := recover(); r != nil {
+			buf := make([]byte, 4096)
+			buf = buf[:runtime.Stack(buf, false)]
+			logger.Criticalf("Unhandled panic: \n%v\n%s", r, buf)
+			os.Exit(exit_panic)
+		}
+	}()
+
+	ctx, err := cmd.DefaultContext()
+	if err != nil {
+		cmd.WriteError(os.Stderr, err)
+		os.Exit(exit_err)
+	}
+
+	var code int
+	commandName := filepath.Base(args[0])
+        code, err = hookToolMain(commandName, ctx, args)
+	if err != nil {
+		cmd.WriteError(ctx.Stderr, err)
+	}
+	return code
+}