@hpidcock 2020-03-12
I do prefer this to one binary rules them all approach. This seems more of a 2.8 chunk of work, so as you mentioned I'd rather this off develop.

Bunch of k8s work would also need to be done, we move the jujud binary around in a few places such as in the dockerfile, init scripts etc. Symlink creation also would need to be updated in k8s. 

One thing that I'm kinda curious about is why jujud startup time is slow. I don't think it would have much to do about the size of the binary (there is some level of caching in the kernel), rather than all the init code over the whole project. Would be good to do some pre-main perf analysis to see where we spend the most time.