@jameinel 2020-03-12
As far as the size of the codebase, import worker/uniter/runner/jujuc and components.RegisterForServer were both necessary to go from 80MB down to 9MB.
I'm guessing all the server code that registers what API/etc are available is taking at least some of the time.
It's pretty hard to profile what happens *before* you hit main().
Supposedly you can use 'perf' but I didn't find a great way to figure it out.