@achilleasa 2020-03-13
I wonder if the jujuc start time could be reduced further by mounting a ramdisk and copying the stripped jujuc binary there.

Also, just curious: as the jujuc server already talks to the unit agent over unix sockets, did we ever investigate daemonizing jujuc and making the hook tools dumb shell scripts that talk to jujuc over another unix socket (passing the relevant envvars as peer creds)?