@hpidcock 2020-03-16
I don't think a ramdisk is necessary. I do like your suggestion of using shell scripts for the jujuc instead, but I think it will bring issues with windows agents, we will need to duplicate scripts with powershell. 

Additionally jujuc communicates over TCP+TLS for k8s. I would actually possibly consider we use http/https for jujuc, curl could then be used. Not sure if curl supports unix socks but worth looking into.