@jameinel 2020-03-16
"jujuc server already talks to the unit agent over unix sockets"
The jujuc *server* runs inside the unit agent. The jujuc client that we invoke from hook scripts connects to that over the unix socket.
So it isn't worth running an intermediary, AFAICT.

I did look at updating charms to speak directly to the agent socket, as the actual structs being passed back and forth are small. Unfortunately it uses 'gob' as the encoding for the structs, which means it is non-trivial to bring into Python. I did find [pygob](https://github.com/mgeisler/pygob). I don't really want to add it as a dependency, but I can prototype with it, I guess.