@jameinel 2020-03-19
Another point that I ran into, while playing with this. 'juju-with-jujuc upgrade-controller --build-agent' will fail if your current GOPATH juju/juju directory doesn't have jujuc (because it is looking to build jujuc and put it in the tarball). conversely 'juju-without upgrade-controller' where your source tree *does* have it, won't complain and will happily upgrade you to a tarball that doesn't have it.
I don't know a concrete way around it, unless we change '--build-agent' to not directly build anything but to call "make install" in the source directory, and have *that* object build the tarball.
(I do still think we should have a clearer 'make tool-tarball' step, rather than multiple places with that sort of logic.)